requirements.txt
amqp==5.3.1
asgiref==3.9.1
beautifulsoup4==4.13.5
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.5
djangorestframework==3.16.1
fake-useragent==2.2.0
greenlet==3.2.4
idna==3.10
kombu==5.5.4
loguru==0.7.3
packaging==25.0
pillow==11.3.0
playwright==1.54.0
prompt_toolkit==3.0.52
pyee==13.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==6.4.0
requests==2.32.5
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
